Split rcpsp_models.py into smaller files
Some care is needed for RCPSPSolution that needs RCPSPModel which in
returns needs RCPSPSolution.
We avoid circular imports by importing rcpsp_model in rcsp_solution only
when type checking. Similarly we avoid importing robust_rcpsp (which
imports rcpsp_model) by checking if the model is a RobustModel instead
of directly checking if it is a AggregRCPSPModel.

We follow pep8 by modifying Aggreg_RCPSPModel into AggregRCPSPModel.
