+++ discrete_optimization/rcpsp/solver/cpm.py
@@ -16,7 +16,8 @@ from discrete_optimization.generic_tools.do_problem import (
 from discrete_optimization.generic_tools.result_storage.result_storage import (
     ResultStorage,
 )
-from discrete_optimization.rcpsp.rcpsp_model import RCPSPModel, RCPSPSolution
+from discrete_optimization.rcpsp.rcpsp_model import RCPSPModel
+from discrete_optimization.rcpsp.rcpsp_solution import RCPSPSolution
 from discrete_optimization.rcpsp.rcpsp_utils import compute_graph_rcpsp
 from discrete_optimization.rcpsp.solver.rcpsp_solver import SolverRCPSP
