+++ tests/rcpsp/solver/test_rcpsp_cp_lns.py
@@ -9,8 +9,9 @@ from discrete_optimization.generic_tools.lns_mip import TrivialInitialSolution
 from discrete_optimization.generic_tools.result_storage.result_storage import (
     from_solutions_to_result_storage,
 )
-from discrete_optimization.rcpsp.rcpsp_model import RCPSPModel, RCPSPSolution
+from discrete_optimization.rcpsp.rcpsp_model import RCPSPModel
 from discrete_optimization.rcpsp.rcpsp_parser import get_data_available, parse_file
+from discrete_optimization.rcpsp.rcpsp_solution import RCPSPSolution
 from discrete_optimization.rcpsp.rcpsp_utils import (
     plot_resource_individual_gantt,
     plot_ressource_view,
