+++ tests/rcpsp/test_rcpsp_model.py
@@ -8,12 +8,9 @@ from typing import Dict, Hashable, Tuple
 
 import pytest
 
-from discrete_optimization.rcpsp.rcpsp_model import (
-    RCPSPModel,
-    RCPSPSolution,
-    TaskDetails,
-)
+from discrete_optimization.rcpsp.rcpsp_model import RCPSPModel
 from discrete_optimization.rcpsp.rcpsp_parser import get_data_available, parse_file
+from discrete_optimization.rcpsp.rcpsp_solution import RCPSPSolution, TaskDetails
 
 files_rcpsp = get_data_available()
 single_modes_files = [f for f in files_rcpsp if "sm" in f]
